requirements.txt
alembic==1.6.5
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
packaging==20.9
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.3.23
toml==0.10.2
urllib3==1.26.6
Werkzeug==2.0.1
